[fix] use INTERN instead of GENTEMP (https://github.com/digikar99/polymorphic-functions/issues/3)
We do want STATIC-DISPATCH-NAME symbol collision to actually take place
when type lists of two polymorphs are "equivalent".

(Credits to phoe for pointing out in the issue) Consider a file A to be
compiled before restarting a lisp image, and file B after the
restart. The use of GENTEMP meant that two "separate" compilations of
the same polymorph in the two files, could result in different
STATIC-DISPATCH-NAMEs. If the two files were then loaded
simultaneously, and the polymorphs static-dispatched at some point,
then there remained the possibility that different static-dispatches
could be using "different versions" of the polymorph.

Thus, we actually do want collisions to take place so that a same
deterministic/latest version of the polymorph is called; therefore we
use INTERN.