[fix] resolve more ambiguous ordering cases of required
Prior to this, the order of (array string) and (string array) was
undefined. Further, this undefined ordering went undetected, in that
this was not caught by type-list-causes-ambiguous-call-p.

Now, such type-lists are resolved according to generic-function
conventions that the first time the classes are different from left
to right, the more specific method has the class as a subclass of the
less specific class. In the case subclass relation does not hold, then
the more specific class has the class earlier in the
class-precedence-list compared to the less specific method. (CONFIRM.)

However, the latter about class-precedence-list is non-sensical in the
case of types because something like type-precedence-list does not
exist. Thus, we are compelled to indicate ambiguous call in these cases.