[revert][change] (array t) should not be a subtype of (array <type>)
This reverts commit b634c81cc9b37d92edd0669b1bf501949de7a28c.

Doing so exponentially increases compilation times on SBCL.
See the following for a relevant issue:

https://sourceforge.net/p/sbcl/mailman/sbcl-help/thread/87sfjdobko.wl-shubhamayare%40yahoo.in/#msg37725365

A motivation for not allowing (array t) to be a subtype of (array <type>)
was to allow for polymorphs specializing on (array <type>), (simple-array <type>),
and (array t) to coexist. But perhaps, a better way might be to implement
appropriate low level functions operating on arrays with element type T,
analogous to arrays with specialized element-types such as single-float, double-float.